edit cmake

+++ include/graph/bipartite_graph.h
@@ -41,7 +41,7 @@ public:
     template <class Container>
     int count(Container const& order) const;
     [[nodiscard]] const std::vector<std::vector<int>>& computeCrossingMatrix() const;
-    int calculateMinimumCrossingLowerBound() const;
+    long long calculateMinimumCrossingLowerBound() const;
 };
 
 #endif
